refactor[#21] : icon shared 레이어로 이동
