+++ beta-reader/src/shared/assets/ArrowUp.tsx
