+++ beta-reader/src/shared/assets/CloseIcon.tsx
