+++ beta-reader/src/shared/assets/MoreIcon.tsx
