+++ beta-reader/src/shared/assets/checkIcon.tsx
