+++ beta-reader/src/shared/assets/index.ts
