+++ beta-reader/src/shared/ui/Form/ui/Checkbox.tsx
@@ -2,7 +2,7 @@ import { cva } from "class-variance-authority";
 import React, { useState } from "react";
 
 import { CheckboxGroupContext, useCheckboxGroupContext } from "../model";
-import { CheckIcon } from "@/public/assets";
+import { CheckIcon } from "@/src/shared/assets";
 
 interface CheckboxGroupProps
   extends React.InputHTMLAttributes<HTMLInputElement> {
